+++ pipelines/data_processing/autorag/documents_indexing_pipeline/tests/pipeline_resource_expectations.py
@@ -1,0 +1,11 @@
+"""Expected Kubernetes CPU/memory tiers for the documents indexing pipeline."""
+
+from kfp_components.utils.pipeline_task_resources import ExecutorResources
+
+WORKLOAD_RESOURCES = ExecutorResources("2", "8Gi", "32", "64Gi")
+
+AUTORAG_INDEXING_EXECUTOR_RESOURCES = {
+    "documents-discovery": WORKLOAD_RESOURCES,
+    "text-extraction": WORKLOAD_RESOURCES,
+    "documents-indexing": WORKLOAD_RESOURCES,
+}
