+++ pipelines/training/automl/autogluon_tabular_training_pipeline/tests/pipeline_resource_expectations.py
@@ -1,0 +1,18 @@
+"""Expected Kubernetes CPU/memory tiers for the tabular training pipeline."""
+
+from kfp_components.utils.pipeline_task_resources import ExecutorResources
+
+STAGE_MAP_RESOURCES = ExecutorResources("0.5", "512Mi", "1", "1Gi")
+WORKLOAD_RESOURCES = ExecutorResources("2", "8Gi", "32", "64Gi")
+LEADERBOARD_RESOURCES = ExecutorResources("1", "4Gi", "32", "64Gi")
+TRAINING_SPEED_RESOURCES = ExecutorResources("4", "16Gi", "32", "64Gi")
+TRAINING_BALANCED_RESOURCES = ExecutorResources("8", "32Gi", "32", "64Gi")
+
+AUTOML_TABULAR_EXECUTOR_RESOURCES = {
+    "publish-component-stage-map": STAGE_MAP_RESOURCES,
+    "automl-data-loader": WORKLOAD_RESOURCES,
+    "autogluon-models-training": TRAINING_BALANCED_RESOURCES,
+    "leaderboard-evaluation": LEADERBOARD_RESOURCES,
+    "autogluon-models-training-2": TRAINING_SPEED_RESOURCES,
+    "leaderboard-evaluation-2": LEADERBOARD_RESOURCES,
+}
