+++ pipelines/training/automl/autogluon_timeseries_training_pipeline/tests/pipeline_resource_expectations.py
@@ -1,0 +1,18 @@
+"""Expected Kubernetes CPU/memory tiers for the time series training pipeline."""
+
+from kfp_components.utils.pipeline_task_resources import ExecutorResources
+
+STAGE_MAP_RESOURCES = ExecutorResources("0.5", "512Mi", "1", "1Gi")
+WORKLOAD_RESOURCES = ExecutorResources("2", "8Gi", "32", "64Gi")
+LEADERBOARD_RESOURCES = ExecutorResources("1", "4Gi", "32", "64Gi")
+TRAINING_SPEED_RESOURCES = ExecutorResources("4", "16Gi", "32", "64Gi")
+TRAINING_BALANCED_RESOURCES = ExecutorResources("8", "32Gi", "32", "64Gi")
+
+AUTOML_TIMESERIES_EXECUTOR_RESOURCES = {
+    "publish-component-stage-map": STAGE_MAP_RESOURCES,
+    "timeseries-data-loader": WORKLOAD_RESOURCES,
+    "autogluon-timeseries-models-training": TRAINING_BALANCED_RESOURCES,
+    "leaderboard-evaluation": LEADERBOARD_RESOURCES,
+    "autogluon-timeseries-models-training-2": TRAINING_SPEED_RESOURCES,
+    "leaderboard-evaluation-2": LEADERBOARD_RESOURCES,
+}
